ARN-2715 Updates drug tests (#834)
