+++ cypress/e2e/v1.0/journeys/common.ts
@@ -1,5 +1,10 @@
 // eslint-disable-next-line import/prefer-default-export
-export const testPractitionerAnalysis = (sectionName: string, origin: string, destination: string) => {
+export const testPractitionerAnalysis = (
+  sectionName: string,
+  origin: string,
+  destination: string,
+  conditionalFlag: boolean,
+) => {
   describe(`Destination: ${destination}`, () => {
     it(`routes to ${destination}`, () => {
       cy.visitStep(origin)
@@ -18,8 +23,9 @@ export const testPractitionerAnalysis = (sectionName: string, origin: string, de
         cy.getQuestion(question).getRadio('No').clickLabel()
       })
 
-      // todo only for Drug use
-      // cy.getQuestion('Does Sam seem motivated to stop or reduce their drug use?').getRadio('Unknown').clickLabel()
+      if (sectionName === 'Drug use' && conditionalFlag) {
+        cy.getQuestion('Does Sam seem motivated to stop or reduce their drug use?').getRadio('Unknown').clickLabel()
+      }
 
       cy.assertResumeUrlIs(sectionName, origin)
       cy.get('#tab_practitioner-analysis').click()
